Move out more functionality
