+++ src/helpers/index.js
@@ -2,5 +2,6 @@ import getElementPosition from './getElementPosition';
 import buildMessage from './buildMessage';
 import getFileContents from './getFileContents';
 import NormalizeOutput from './normalizeOutput.js';
+import CreateReportsJson from './createReportsJson.js';
 
-export { getElementPosition, buildMessage, getFileContents, NormalizeOutput };
+export { getElementPosition, buildMessage, getFileContents, NormalizeOutput, CreateReportsJson };
