+++ src/logger.js
@@ -90,7 +90,7 @@ logger.generalError = (error: string): string => {
 
   console.error(chalk.red(error));
 
-  return chalk.red(error);
+  return error;
 };
 
 logger.log = (message: string): string => {
